requirements.txt
fastapi[all]
python-dotenv
uvicorn
jinja2
openai
httpx
python-multipart
